server/requirements.txt
mysql-connector-python>=8.0
Flask>=1.1
flask_cors>=3.0.7
python-dotenv>=0.10.3
